+++ StoryEditor/out/pre.js
@@ -1,0 +1,4 @@
+// let serverURL = "http://localhost:3000";
+let serverURL = "https://shocking-button-sidecar.glitch.me/";
+const AID = "__SE-"; // app id
+//# sourceMappingURL=pre.js.map