+++ StoryEditor_1/out/main.js
@@ -1,54 +1,3 @@
-let b_theme = document.querySelector(".b-theme");
-let themeImg = b_theme.querySelector("img");
-let theme = localStorage.getItem("__SE-theme") || "light";
-let i_title = pane_editBoard.querySelector(".i-title");
-let ta_text = pane_editBoard.querySelector(".ta-text");
-let _editBoard_b;
-function setTheme(val) {
-    let html = document.body.parentElement;
-    // html.style.filter = (val == "dark" ? "invert(1) hue-rotate(180deg) brightness(0.8) contrast(0.8)" : "none");
-    html.style.setProperty("--theme-filter", (val == "dark" ? "invert(1) hue-rotate(180deg) brightness(0.8) contrast(0.8)" : "none"));
-    html.style.setProperty("--img-filter", (val == "dark" ? "invert(1) hue-rotate(180deg) brightness(1) contrast(1)" : "none"));
-    localStorage.setItem("__SE-theme", val);
-    themeImg.src = (val == "dark" ? "assets/dark_mode.svg" : "assets/light_mode.svg");
-}
-setTheme(theme);
-b_theme.addEventListener("click", e => {
-    if (theme == "dark")
-        theme = "light";
-    else
-        theme = "dark";
-    setTheme(theme);
-});
-let b_open = document.querySelector(".b-open");
-let b_create = document.querySelector(".b-create");
-async function openDir() {
-    story.handle = await showDirectoryPicker({
-        mode: "readwrite",
-        id: "openDir"
-    });
-}
-b_open.addEventListener("click", e => {
-    // openDir();
-    // let name = prompt("Please type the name of the project you wish to join");
-    openProjectMenu();
-});
-b_create.addEventListener("click", e => {
-    let name = prompt("Please enter the project name:");
-    if (!name)
-        return;
-    let code = prompt("Please enter a passcode in order to edit the project:");
-    if (!code)
-        return;
-    socket.emit("createProject", name, code, (res) => {
-        if (res == 1) {
-            alert("Successfully created project: " + name);
-            localStorage.setItem(AID + "code-" + name, code);
-        }
-        else
-            alert("Failed to create project");
-    });
-});
 let i_searchAll = document.querySelector(".i-search-all");
 let choiceList = document.querySelector(".choice-list");
 let i_addChoice = document.querySelector(".i-add-choice");
@@ -60,27 +9,6 @@ let b_playFromHere = document.querySelector(".b-play-from-here");
 let b_export = document.querySelector(".b-export");
 let b_import = document.querySelector(".b-import");
 let b_reset = document.querySelector(".b-reset");
-const b_chooseBGImg = document.querySelector(".b-choose-bg-img");
-const b_removeBGImg = document.querySelector(".b-remove-bg-img");
-const l_bgPreview = document.querySelector(".l-bg-preview");
-const img_bgPreview = document.querySelector(".img-bg-preview");
-b_chooseBGImg.addEventListener("click", async (e) => {
-    if (story.selBoards.length != 1)
-        return;
-    let sel = story.selBoards[0];
-    let name = await chooseImage();
-    if (!name)
-        return;
-    sel.setImg(name);
-});
-b_removeBGImg.addEventListener("click", async (e) => {
-    if (story.selBoards.length != 1)
-        return;
-    let sel = story.selBoards[0];
-    if (sel.img == null)
-        return;
-    sel.setImg(null);
-});
 b_export.addEventListener("click", e => {
     let o = story.getSaveObj();
     if (!o)
@@ -133,44 +61,20 @@ b_addChoice.addEventListener("click", e => {
     let name = i_addChoice.value;
     if (!name)
         return;
-    let list = name.split(",");
-    let res = sel.addChoice(list);
-    socket.emit("s_addChoice", sel._id, list);
+    let res = sel.addChoice(name.split(","));
     story.save();
     i_addChoice.value = "";
     // loadEditBoard(sel);
     story.selectBoard(res[0]);
 });
 b_play.addEventListener("click", e => {
-    if (!story)
-        return;
     story._save();
     setPlayI(0);
-    // location.pathname = `/play/index.html?email=${story.owner}&pid=${story.filename}`;
-    let url = new URL(location.href);
-    url.pathname = "play/";
-    url.searchParams.set("email", story.owner);
-    url.searchParams.set("pid", story.filename);
-    // location.assign(url);
-    let a = document.createElement("a");
-    a.href = url.href;
-    a.target = "_blank";
-    a.click();
+    location.pathname += "play";
 });
 b_resumePlay.addEventListener("click", e => {
-    if (!story)
-        return;
     story._save();
-    // location.pathname = `/play/index.html?email=${story.owner}&pid=${story.filename}`;
-    let url = new URL(location.href);
-    url.pathname = "play/";
-    url.searchParams.set("email", story.owner);
-    url.searchParams.set("pid", story.filename);
-    // location.assign(url);
-    let a = document.createElement("a");
-    a.href = url.href;
-    a.target = "_blank";
-    a.click();
+    location.pathname += "play";
 });
 b_save.addEventListener("click", e => {
     story._save();
@@ -179,8 +83,6 @@ function simplifyText(t) {
     return t.toLowerCase().replaceAll(" ", "");
 }
 i_searchAll.addEventListener("input", e => {
-    if (!story)
-        return;
     let v = simplifyText(i_searchAll.value);
     for (const b of story.loadedObjs) {
         let a = simplifyText(b.title);
@@ -211,8 +113,7 @@ function initPane(c) {
 }
 function closePane(c) {
     c.classList.remove("open");
-    if (c == pane_editBoard)
-        _editBoard_b = null;
+    console.log("CLSOE");
 }
 function closeAllPanes() {
     for (const c of panes) {
@@ -234,8 +135,6 @@ function getM(e) {
     return { x, y };
 }
 document.addEventListener("mousemove", e => {
-    if (!story)
-        return;
     let { x, y } = getM(e);
     let dx = x - story.lx;
     let dy = y - story.ly;
@@ -245,16 +144,26 @@ document.addEventListener("mousemove", e => {
         story.setPan(story.panX - dx, story.panY - dy);
     }
     else if (story.dragBoards.length) {
-        story.moveBoards(story.dragBoards, dx, dy);
+        let done = [];
+        for (const b of story.dragBoards) {
+            function loop(board, once = false) {
+                if (done.includes(board))
+                    return;
+                board.x += dx;
+                board.y += dy;
+                board.update();
+                done.push(board);
+                if (once)
+                    return;
+                for (const c of board.buttons) {
+                    loop(c.board);
+                }
+            }
+            loop(b, keys.alt);
+        }
     }
 });
-let mouseDown = [false, false, false];
 document.addEventListener("mousedown", e => {
-    if (menus.children.length)
-        return;
-    mouseDown[e.button] = true;
-    if (!story)
-        return;
     if (e.clientY <= 72)
         return;
     i_searchAll.value = "";
@@ -284,7 +193,6 @@ document.addEventListener("mousedown", e => {
                 if (cancel)
                     return;
                 par.addChoice(list, children);
-                socket.emit("s_addChoice", par._id, list, children.map(v => v._id));
                 story.save();
                 story.deselectBoards();
                 children.forEach(v => story.selectAddBoard(v));
@@ -296,14 +204,11 @@ document.addEventListener("mousedown", e => {
             story.dragBoards = [story.hoverBoard];
         // story.selectBoard(story.hoverBoard);
     }
-    else if (!overPane && menus.children.length == 0) {
+    else if (!overPane) {
         story.isPanning = true;
     }
 });
 document.addEventListener("mouseup", e => {
-    mouseDown[e.button] = false;
-    if (!story)
-        return;
     if (!story.hoverBoard)
         if (!overPane)
             if (story.sx == story.lx && story.sy == story.ly) {
@@ -377,26 +282,21 @@ class HistState {
     label;
 }
 let story;
-if (false) {
-    if (localStorage.getItem("__SELS-tmp")) {
-        story = Story.load();
-        story.origin.load();
-        story.makeConnection(story.origin, story.start, ConnectionType.start);
-    }
-    else {
-        story = new Story("TestFile1", "");
-        story.init();
-    }
+if (localStorage.getItem("__SELS-tmp")) {
+    story = Story.load();
+    story.origin.load();
+    story.makeConnection(story.origin, story.start, ConnectionType.start);
 }
-initNetworkFromEditor();
+else {
+    story = new Story("TestFile1");
+    story.init();
+}
 function resetFile() {
     localStorage.removeItem("__SELS-tmp");
     location.reload();
 }
 // Auto Save Interval
 setInterval(() => {
-    if (!story)
-        return;
     if (story.needsSave) {
         story._save();
         story.needsSave = false;