+++ StoryEditor_1/out/play.js
@@ -37,6 +37,13 @@ let playStory;
 let allBoards = [];
 let passage = document.querySelector(".passage");
 // let buttons = document.querySelector(".buttons");
+function wait(delay) {
+    return new Promise(resolve => {
+        setTimeout(() => {
+            resolve();
+        }, delay);
+    });
+}
 function madeChoice(btn, button) {
     let p_btns = document.querySelectorAll(".p-btn");
     for (const c of p_btns) {
@@ -51,35 +58,13 @@ function scrollDown() {
 async function loadBoard(b) {
     playData.locId = b._id;
     playData.save();
-    let cont = document.createElement("div");
-    passage.appendChild(cont);
-    let imgUrl = b.img;
-    console.log("img", imgUrl);
-    if (imgUrl) {
-        let img = document.createElement("img");
-        img.src = `${serverURL}/projects/${playStory.owner}/${playStory.filename}/images/${imgUrl}`;
-        let success = false;
-        await new Promise(resolve => {
-            img.onload = function () {
-                success = true;
-                resolve();
-            };
-            img.onerror = function () {
-                resolve();
-            };
-        });
-        // if(success){
-        cont.appendChild(img);
-        await wait(4000);
-        // }
-    }
     let lines = b.text.replaceAll("\n", " ").split(". ").map(v => v.trim());
     await wait(500);
     for (const l of lines) {
         let div = document.createElement("div");
         div.className = "p-item";
         div.textContent = l;
-        cont.appendChild(div);
+        passage.appendChild(div);
         scrollDown();
         await wait(2000);
     }
@@ -87,13 +72,13 @@ async function loadBoard(b) {
         let end = document.createElement("div");
         end.className = "end";
         end.textContent = "END";
-        cont.appendChild(end);
+        passage.appendChild(end);
         scrollDown();
         return;
     }
     await wait(500);
     let buttons = document.createElement("div");
-    cont.appendChild(buttons);
+    passage.appendChild(buttons);
     buttons.className = "buttons";
     for (const btn of b.buttons) {
         let button = document.createElement("button");
@@ -113,65 +98,9 @@ async function loadBoard(b) {
         await wait(500);
     }
 }
-async function initPlay() {
-    let url = new URL(location.href);
-    let email = url.searchParams.get("email");
-    let pid = url.searchParams.get("pid");
-    if (!email || !pid) {
-        console.warn("Invalid search params");
-        return;
-    }
-    console.log("...starting load");
-    // 
-    let code = LSGet("code-" + pid);
-    let pdata;
-    async function promptCode(isFirst = false) {
-        if (!isFirst) {
-            code = prompt("Please enter project pass code:");
-            if (code == null)
-                return;
-        }
-        pdata = await new Promise(resolve => {
-            socket.emit("openProject_readonly", email, pid, code, ((data) => {
-                resolve(data);
-            }));
-        });
-        if (!pdata) {
-            console.log("could not find pdata");
-            return;
-        }
-        if (pdata.err) {
-            if (pdata.code == 0) {
-                alert(pdata.err);
-                return;
-            }
-            else if (pdata.code == 1) {
-                alert(pdata.err);
-                await promptCode();
-            }
-        }
-    }
-    console.log("...finding code");
-    await promptCode(true);
-    console.log("...found?");
-    localStorage.setItem(AID + "code-" + pid, code || "");
-    if (!pdata) {
-        console.warn("could not get pdata");
-        alert("Failed to find/load project");
-        return;
-    }
-    console.warn("OWNER:", pdata.owner);
-    if (!pdata.owner) {
-        console.warn("no owner found");
-        return;
-    }
-    pdata.storyData.owner = pdata.owner;
-    story = Story.load(pdata.storyData);
-    console.log("story", story);
-    // 
+function initPlay() {
     playData = PlayData.load();
-    // playStory = Story.load();
-    playStory = story;
+    playStory = Story.load();
     for (const b of playStory.loadedObjs) {
         if (b instanceof Board) {
             allBoards[b._id] = b;
@@ -183,13 +112,7 @@ async function initPlay() {
         playData.locId = playStory.start._id;
         start = allBoards[playData.locId];
     }
-    loadBoard(start);
+    loadBoard(allBoards[playData.locId]);
 }
-// @ts-ignore
-let myCursor;
-// @ts-ignore
-let socket = io(serverURL);
-socket.on("connect", () => {
-    initPlay();
-});
+initPlay();
 //# sourceMappingURL=play.js.map