+++ StoryEditor/changelog.md
@@ -1,9 +1,31 @@
-### 1/17/23
+### 1/17/24
 - initial creation
 - crazy amount of stuff
 
-### 1/18/23
+### 1/18/24
 - also massive update
 - so many bug fixes
 - play mode
 - flow arrows
+
+### 1/19/24
+- theme system with button toggle in header
+- start of server setup
+
+### 1/20/24
+- fixed dates to be 2024 xD
+
+### 1/21/24
+- multiplayer system
+- other stuff probably
+- rediculously massive update
+
+### 1/22/24
+- mostly finished up
+- fixed lots and lots of bugs
+- added multiplayer support for reconnections
+- added image support
+- added play mode online support
+- added create projects
+- added looking through open projects
+- added pass codes to projects